Fix edge case when margins are applied to scroll items.
This bug was pre-existing and was not caused by changes in the
last six commits.

The Bug / Edge Case
-------------------
When scroll items have margins the top margin of the first item
and the bottom margin of the last item count towards the 'scrollTop'
and 'scrollLeft' but, due to collapsing margins, do not count towards
the height of the inner-container (formally the 'document'). Thus, when fully scrolled to the bottom (or right), the 'viewPortTop' is larger than the
'distanceToViewportTop' (by double the margin on the scroll items).
When scrolling quickly the situation when the two are exactly equal (i.e. when viewPortTop - distanceToViewportTop === 0 ) is often skipped leading
to a failure in the infinite scroll.

The Fix
--------
This has been rectified by testing for when the viewPortTop is greater than
the distanceToViewPortTop.